refactor-ir-cake: SizeOf, CostOf removed from IR + more ScalaDocs
