+++ sc/shared/src/main/scala/sigma/compiler/ir/Modules.scala
@@ -2,6 +2,9 @@ package sigma.compiler.ir
 
 import sigma.compiler.ir.meta.ModuleInfo
 
+/** Defines registration of IR modules.
+  * @see ModuleInfo
+  */
 trait Modules extends Base { self: IRContext =>
 
   /** Whether staged modules should be registered when cake is constructed and initialized. */
