+++ sc/shared/src/main/scala/sigma/compiler/ir/primitives/Equal.scala
@@ -4,6 +4,7 @@ import sigma.compiler.ir.{Base, IRContext}
 
 import scala.annotation.unused
 
+/** Defines IR representation of equality operations. */
 trait Equal extends Base { self: IRContext =>
   /** Binary operation representing structural equality between arguments. */
   case class Equals[A: Elem]() extends BinOp[A, Boolean]("==") {
